All "secondary" was renamed to "remote"

+++ test/suites/profiling.bash
@@ -7,9 +7,12 @@ normalize_gcno_file() {
 
 
 SUITE_profiling_PROBE() {
-    touch test.c
+    echo 'int main(void) { return 0; }' >test.c
     if ! $COMPILER -fprofile-generate -c test.c 2>/dev/null; then
         echo "compiler does not support profiling"
+    fi
+    if ! $COMPILER -fprofile-generate test.o -o test 2>/dev/null; then
+        echo "compiler cannot link with profiling"
     fi
     if ! $COMPILER -fprofile-generate=data -c test.c 2>/dev/null; then
         echo "compiler does not support -fprofile-generate=path"
@@ -50,6 +53,7 @@ SUITE_profiling() {
     expect_stat no_input_file 1
 
     # -------------------------------------------------------------------------
+if $RUN_WIN_XFAIL; then
     TEST "-fprofile-use"
 
     $CCACHE_COMPILE -fprofile-generate -c test.c
@@ -75,8 +79,10 @@ SUITE_profiling() {
     $CCACHE_COMPILE -fprofile-use -c test.c
     expect_stat direct_cache_hit 1
     expect_stat cache_miss 3
+fi
 
     # -------------------------------------------------------------------------
+if $RUN_WIN_XFAIL; then
     TEST "-fprofile-use=dir"
 
     mkdir data
@@ -104,8 +110,9 @@ SUITE_profiling() {
     $CCACHE_COMPILE -fprofile-use=data -c test.c
     expect_stat direct_cache_hit 1
     expect_stat cache_miss 3
-
+fi
     # -------------------------------------------------------------------------
+if $RUN_WIN_XFAIL; then
     TEST "-fprofile-generate=dir in different directories"
 
     mkdir -p dir1/data dir2/data
@@ -155,38 +162,38 @@ SUITE_profiling() {
         || test_failed "compilation error"
     # Note: No expect_stat here since GCC and Clang behave differently – just
     # check that the compiler doesn't warn about not finding the profile data.
-
+fi
     # -------------------------------------------------------------------------
-    TEST "-ftest-coverage with -fprofile-dir"
+    if $COMPILER_TYPE_GCC; then
+        # GCC 9 and newer creates a mangled .gcno filename (still in the current
+        # working directory) if -fprofile-dir is given.
+        for flag in "" -fprofile-dir=dir; do
+            for dir in . subdir; do
+                TEST "-ftest-coverage with -fprofile-dir=$flag, dir=$dir"
+                $CCACHE -z >/dev/null
 
-    # GCC 9 and newer creates a mangled .gcno filename (still in the current
-    # working directory) if -fprofile-dir is given.
+                mkdir -p "$dir"
+                touch "$dir/test.c"
+                find -name '*.gcno' -delete
 
-    for flag in "" -fprofile-dir=dir; do
-        for dir in . subdir; do
-            $CCACHE -z >/dev/null
+                $COMPILER $flag -ftest-coverage -c $dir/test.c -o $dir/test.o
+                gcno_name=$(find -name '*.gcno')
+                rm "$gcno_name"
 
-            mkdir -p "$dir"
-            touch "$dir/test.c"
-            find -name '*.gcno' -delete
+                $CCACHE_COMPILE $flag -ftest-coverage -c $dir/test.c -o $dir/test.o
+                expect_stat direct_cache_hit 0
+                expect_stat cache_miss 1
+                expect_exists "$gcno_name"
+                rm "$gcno_name"
 
-            $COMPILER $flag -ftest-coverage -c $dir/test.c -o $dir/test.o
-            gcno_name=$(find -name '*.gcno')
-            rm "$gcno_name"
-
-            $CCACHE_COMPILE $flag -ftest-coverage -c $dir/test.c -o $dir/test.o
-            expect_stat direct_cache_hit 0
-            expect_stat cache_miss 1
-            expect_exists "$gcno_name"
-            rm "$gcno_name"
-
-            $CCACHE_COMPILE $flag -ftest-coverage -c $dir/test.c -o $dir/test.o
-            expect_stat direct_cache_hit 1
-            expect_stat cache_miss 1
-            expect_exists "$gcno_name"
-            rm "$gcno_name"
+                $CCACHE_COMPILE $flag -ftest-coverage -c $dir/test.c -o $dir/test.o
+                expect_stat direct_cache_hit 1
+                expect_stat cache_miss 1
+                expect_exists "$gcno_name"
+                rm "$gcno_name"
+            done
         done
-    done
+    fi
 
     # -------------------------------------------------------------------------
     TEST "-fprofile-arcs for different object file paths"
